feat(TAT-138): reorganize files
