+++ src/router/index.ts
@@ -1,8 +1,8 @@
 import { createRouter, createWebHistory } from "vue-router";
-import Home from "@/components/Home.vue";
-import Methodology from "@/components/Methodology.vue";
-import Help from "@/components/Help.vue";
-import Calculator from "@/components/Calculator.vue";
+import Home from "@/views/Home.vue";
+import Methodology from "@/views/Methodology.vue";
+import Help from "@/views/Help.vue";
+import Calculator from "@/views/Calculator.vue";
 
 const routes = [
   {
