+++ src/views/Home.vue
@@ -1,22 +1,18 @@
 <template>
   <div>
     <div>This is the home page</div>
-    <div>Data from store: {{ exampleStore.name }}</div>
   </div>
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import { useExampleStore } from '../stores/example.store';
 
 export default defineComponent({
   data() {
     return {
-      exampleStore: useExampleStore(),
     }
   }
 })
 </script>
 
-<style scoped>
-</style>
+<style scoped></style>
