+++ src/views/Methodology.vue
